separate CognitiveLoad into separate sub-package
