+++ src/cz/cvut/cognitive/distractors/CollectObjectDistraction.java
@@ -1,28 +1,22 @@
 
 package cz.cvut.cognitive.distractors;
 
-import com.bulletphysics.collision.shapes.BoxShape;
+import cz.cvut.cognitive.load.CognitiveFunction;
 import com.jme3.asset.AssetManager;
 import com.jme3.asset.TextureKey;
-import com.jme3.bullet.BulletAppState;
-import com.jme3.bullet.collision.shapes.CapsuleCollisionShape;
 import com.jme3.bullet.collision.shapes.CollisionShape;
 import com.jme3.bullet.control.GhostControl;
-import com.jme3.bullet.control.RigidBodyControl;
 import com.jme3.bullet.util.CollisionShapeFactory;
 import com.jme3.collision.CollisionResults;
 import com.jme3.font.BitmapText;
 import com.jme3.material.Material;
 import com.jme3.math.ColorRGBA;
-import com.jme3.math.Matrix3f;
 import com.jme3.math.Ray;
 import com.jme3.math.Vector3f;
 import com.jme3.renderer.Camera;
 import com.jme3.scene.Geometry;
 import com.jme3.scene.Spatial;
-import com.jme3.scene.shape.Box;
 import com.jme3.scene.shape.Sphere;
-import static com.jme3.shader.VarType.Vector3;
 import com.jme3.texture.Texture;
 import eu.opends.car.SteeringCar;
 import eu.opends.main.Simulator;
