+++ src/eu/opends/niftyGui/DistractionScreenGUIController.java
@@ -19,7 +19,7 @@
 package eu.opends.niftyGui;
 
 
-import cz.cvut.cognitive.distractors.CognitiveFunction;
+import cz.cvut.cognitive.load.CognitiveFunction;
 import de.lessvoid.nifty.Nifty;
 import de.lessvoid.nifty.NiftyEventSubscriber;
 import de.lessvoid.nifty.controls.CheckBox;
